Add middleware and JWT auth template with Admin related files

+++ FLA/routes/api.php
@@ -7,12 +7,12 @@ use App\Http\Controllers\playerController;
 use App\Http\Controllers\teamController;
 use Illuminate\Support\Facades\Route;
 
-Route::get('/fetch-live', [ApiController::class, 'testLiveAPI']);
+Route::get('/fetch-live', [ApiController::class, 'testLiveAPI'])->middleware('check_role:admin');
 Route::get('/all-leagues', [leagueController::class, 'getAllLeagues']);
 Route::get('/leagues/{id}', [leagueController::class, 'getLeagueByApiId']);
-Route::post('/leagues', [leagueController::class, 'createLeague']);
-Route::put('/leagues/{id}', [leagueController::class, 'updateLeague']);
-Route::delete('/leagues/{id}', [leagueController::class, 'deleteLeague']);
+Route::post('/leagues', [leagueController::class, 'createLeague']) ->middleware('check_role:admin');
+Route::put('/leagues/{id}', [leagueController::class, 'updateLeague'])->middleware('check_role:admin');
+Route::delete('/leagues/{id}', [leagueController::class, 'deleteLeague'])->middleware('check_role:admin');
 Route::get('/league-table/{league_id}/{season_id}', [leaguePositionController::class, 'getLeagueTable']);
 Route::get('/team-position/{team_id}/{league_id}/{season_id}', [leaguePositionController::class, 'getTeamPosInLeague']);
 Route::get('/all-teams', [teamController::class, 'getAllTeams']);
